requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
google-api-python-client==2.149.0
google-auth==2.35.0
requests==2.32.3
fastapi==0.115.0
uvicorn[standard]==0.30.6

# NOTE: Latest available versions (consider upgrading):
# python-telegram-bot==22.6
# google-api-python-client==2.191.0
# google-auth==2.48.0
# requests==2.32.5
# fastapi==0.135.1
# uvicorn==0.41.0
